// HealthCareApiServer/WebApp/client-app/src/app/api/models/doctor-model.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EntityModel } from '../models/entity-model';
export type DoctorModel = EntityModel & {
'userId': number;
'firstName'?: string | null;
'lastName'?: string | null;
'hospitalId': number;
'hospitalName'?: string | null;
'departmentId': number;
'departmentName'?: string | null;
};